+++ CHANGELOG.md
@@ -73,3 +73,8 @@
 - fix KeyBindOptions not being saved
     - since this is a critical fix, this will also be released for 1.19.2
 - add 'Snap-Perspective' mode to Freelook
+
+### 2.2.8
+
+- HitColor option
+- Allow Snap Perspective mode to be used on servers where Freelook is disallowed